Move websocket proxy to http1.
