+++ rust-proxy/src/health_check/health_check_task.rs
@@ -1,6 +1,6 @@
 use crate::configuration_service::app_config_service::GLOBAL_CONFIG_MAPPING;
 use crate::constants::common_constants::TIMER_WAIT_SECONDS;
-use crate::proxy::http_client::HttpClients;
+use crate::proxy::http1::http_client::HttpClients;
 use crate::vojo::app_config::Route;
 use crate::vojo::health_check::HealthCheckType;
 use crate::vojo::health_check::HttpHealthCheckParam;
