+++ rust-proxy/src/proxy/http1/tls_acceptor.rs
@@ -1,4 +1,4 @@
-use crate::proxy::tls_stream::TlsStream;
+use crate::proxy::http1::tls_stream::TlsStream;
 use core::task::{Context, Poll};
 use futures_util::ready;
 use hyper::server::accept::Accept;
@@ -40,7 +40,7 @@ impl Accept for TlsAcceptor {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::proxy::tls_acceptor::TlsAcceptor;
+    use crate::proxy::http1::tls_acceptor::TlsAcceptor;
     use hyper::service::{make_service_fn, service_fn};
     use hyper::{Body, Response, Server};
     use lazy_static::lazy_static;
