+++ rust-proxy/src/proxy/http2/mod.rs
@@ -1,0 +1,1 @@
+pub mod grpc_proxy;
