+++ rust-proxy/src/vojo/app_config.rs
@@ -55,10 +55,10 @@ impl Route {
                 item.prefix.insert(0, '/')
             }
             let path_rewrite = item.prefix_rewrite.clone();
-            if !path_rewrite.ends_with('/') {
-                let src_prefix_rewrite_len = item.prefix_rewrite.len();
-                item.prefix_rewrite.insert(src_prefix_rewrite_len, '/');
-            }
+            // if !path_rewrite.ends_with('/') {
+            //     let src_prefix_rewrite_len = item.prefix_rewrite.len();
+            //     item.prefix_rewrite.insert(src_prefix_rewrite_len, '/');
+            // }
             if !path_rewrite.starts_with('/') {
                 item.prefix_rewrite.insert(0, '/')
             }
@@ -183,10 +183,8 @@ pub enum ServiceType {
     Http,
     Https,
     Tcp,
-    Websocket,
-    WebsocketTls,
-    Grpc,
-    GrpcTls,
+    Http2,
+    Http2Tls,
 }
 #[derive(Debug, Clone, Default)]
 pub struct ServiceConfig {
